a little less lost
