+++ index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Document</title>
     <link rel="stylesheet" href="./css/style.css" />
-    <script type="module" src="./js/script.js"></script>
+    <script type="module" src="./js/controller.js"></script>
   </head>
   <body>
     <header class="header-container">
