+++ js/helpers/globalFunctions.js
