--- js/script.js
@@ -1,7 +1,0 @@
-import initModel from "./model.js";
-import initView from "./views/views.js";
-import initGlobalFunction from "./globalFunctions.js";
-
-initModel();
-initView();
-initGlobalFunction();
