+++ js/views/productsViews/selectionView.js
@@ -1,5 +1,5 @@
 import * as model from "../../model.js";
-import * as globalFunctions from "../../globalFunctions.js";
+import * as globalFunctions from "../../helpers/globalFunctions.js";
 
 export default function selectProduct() {}
